+++ src/mainwindow.cpp
@@ -18,8 +18,6 @@
  *
  ***********************************************************************/
 
-#include <QDebug>
-
 #include <QApplication>
 #include <QClipboard>
 #include <QCommonStyle>
@@ -38,8 +36,10 @@
 #include <QMenuBar>
 #include <QScrollBar>
 #include <QSettings>
+#include <QSizePolicy>
 #include <QStatusBar>
 #include <QTemporaryFile>
+#include <QTextDocumentFragment>
 
 #include "3rdparty/QtAwesome/QtAwesome.h"
 
@@ -56,7 +56,8 @@
 #include "simplefontdialog.h"
 #include "stylesheetbuilder.h"
 #include "themeselectiondialog.h"
-#include "spelling/dictionary_manager.h"
+#include "spelling/dictionarymanager.h"
+#include "spelling/spellcheckdecorator.h"
 
 namespace ghostwriter
 {
@@ -71,14 +72,15 @@ enum SidebarTabIndex {
 
 #define GW_MAIN_WINDOW_GEOMETRY_KEY "Window/mainWindowGeometry"
 #define GW_MAIN_WINDOW_STATE_KEY "Window/mainWindowState"
-#define GW_SIDEBAR_STATE_KEY "Window/sidebarGeometry"
+#define GW_SPLITTER_GEOMETRY_KEY "Window/splitterGeometry"
 
 MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     : QMainWindow(parent)
 {
+    QString fileToOpen;
+    this->focusModeEnabled = false;
     this->awesome = new QtAwesome(qApp);
     this->awesome->initFontAwesome();
-    QString fileToOpen;
     setWindowIcon(QIcon(":/resources/images/ghostwriter.svg"));
     this->setObjectName("mainWindow");
     this->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Preferred);
@@ -93,6 +95,7 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     MarkdownDocument *document = new MarkdownDocument();
 
     editor = new MarkdownEditor(document, theme.lightColorScheme(), this);
+    editor->setMinimumWidth(0.1 * qApp->primaryScreen()->size().width());
     editor->setFont(appSettings->editorFont().family(), appSettings->editorFont().pointSize());
     editor->setUseUnderlineForEmphasis(appSettings->useUnderlineForEmphasis());
     editor->setEnableLargeHeadingSizes(appSettings->largeHeadingSizesEnabled());
@@ -101,13 +104,16 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     editor->setPlainText("");
     editor->setEditorWidth((EditorWidth) appSettings->editorWidth());
     editor->setEditorCorners((InterfaceStyle) appSettings->interfaceStyle());
-    editor->setSpellCheckEnabled(appSettings->liveSpellCheckEnabled());
     editor->setItalicizeBlockquotes(appSettings->italicizeBlockquotes());
     editor->setTabulationWidth(appSettings->tabWidth());
     editor->setInsertSpacesForTabs(appSettings->insertSpacesForTabsEnabled());
-    connect(editor, SIGNAL(fontSizeChanged(int)), this, SLOT(onFontSizeChanged(int)));
-    this->setFocusProxy(editor);
 
+    connect(editor,
+        &MarkdownEditor::fontSizeChanged,
+        this,
+        &MainWindow::onFontSizeChanged
+    );
+    this->setFocusProxy(editor);
 
     // We need to set an empty style for the editor's scrollbar in order for the
     // scrollbar CSS stylesheet to take full effect.  Otherwise, the scrollbar's
@@ -116,6 +122,9 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     //
     editor->verticalScrollBar()->setStyle(new QCommonStyle());
     editor->horizontalScrollBar()->setStyle(new QCommonStyle());
+
+    spelling = new SpellCheckDecorator(editor);
+    spelling->setLiveSpellCheckEnabled(appSettings->liveSpellCheckEnabled());
 
     buildSidebar();
 
@@ -153,28 +162,9 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
         recentFiles = history.recentFiles(MAX_RECENT_FILES + 2);
     }
 
-    bool fileLoadError = false;
-
     if (!filePath.isNull() && !filePath.isEmpty()) {
-        QFileInfo cliFileInfo(filePath);
-
-        if (!cliFileInfo.exists()) {
-            QFile cliFile(filePath);
-
-            // Try to create a new file if the specified file does not exist.
-            cliFile.open(QIODevice::WriteOnly);
-            cliFile.close();
-
-            if (!cliFile.exists()) {
-                fileLoadError = true;
-                qCritical("Could not create new file. Check permissions.");
-            }
-        }
-
-        if (!fileLoadError) {
-            fileToOpen = filePath;
-            recentFiles.removeAll(cliFileInfo.absoluteFilePath());
-        }
+        fileToOpen = filePath;
+        recentFiles.removeAll(QFileInfo(filePath).absoluteFilePath());
     }
 
     if (fileToOpen.isNull()
@@ -186,7 +176,7 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
             lastFile = recentFiles.first();
         }
 
-        if (QFileInfo(lastFile).exists()) {
+        if (QFileInfo::exists(lastFile)) {
             fileToOpen = lastFile;
             recentFiles.removeAll(lastFile);
         }
@@ -199,6 +189,7 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
         (
             recentFilesActions[i],
             &QAction::triggered,
+            this,
             [this, i]() {
 
                 if (nullptr != recentFilesActions[i]) {
@@ -255,8 +246,8 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     connect(appSettings, SIGNAL(hideMenuBarInFullScreenChanged(bool)), this, SLOT(toggleHideMenuBarInFullScreen(bool)));
     connect(appSettings, SIGNAL(fileHistoryChanged(bool)), this, SLOT(toggleFileHistoryEnabled(bool)));
     connect(appSettings, SIGNAL(displayTimeInFullScreenChanged(bool)), this, SLOT(toggleDisplayTimeInFullScreen(bool)));
-    connect(appSettings, SIGNAL(dictionaryLanguageChanged(QString)), editor, SLOT(setDictionary(QString)));
-    connect(appSettings, SIGNAL(liveSpellCheckChanged(bool)), editor, SLOT(setSpellCheckEnabled(bool)));
+    connect(appSettings, SIGNAL(dictionaryLanguageChanged(QString)), spelling, SLOT(setDictionary(QString)));
+    connect(appSettings, SIGNAL(liveSpellCheckChanged(bool)), spelling, SLOT(setSpellCheckEnabled(bool)));
     connect(appSettings, SIGNAL(editorWidthChanged(EditorWidth)), this, SLOT(changeEditorWidth(EditorWidth)));
     connect(appSettings, SIGNAL(interfaceStyleChanged(InterfaceStyle)), this, SLOT(changeInterfaceStyle(InterfaceStyle)));
     connect(appSettings, SIGNAL(previewTextFontChanged(QFont)), this, SLOT(applyTheme()));
@@ -271,10 +262,10 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
 
     // If we have an available dictionary, then set up spell checking.
     if (!language.isNull() && !language.isEmpty()) {
-        editor->setDictionary(language);
-        editor->setSpellCheckEnabled(appSettings->liveSpellCheckEnabled());
+        spelling->setDictionary(language);
+        spelling->setLiveSpellCheckEnabled(appSettings->liveSpellCheckEnabled());
     } else {
-        editor->setSpellCheckEnabled(false);
+        spelling->setLiveSpellCheckEnabled(false);
     }
 
     this->connect
@@ -310,16 +301,9 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
     connect(outlineWidget, SIGNAL(headingNumberNavigated(int)), htmlPreview, SLOT(navigateToHeading(int)));
     connect(appSettings, SIGNAL(currentHtmlExporterChanged(Exporter *)), htmlPreview, SLOT(setHtmlExporter(Exporter *)));
 
-    htmlPreview->setMinimumWidth(0);
+    htmlPreview->setMinimumWidth(0.1 * qApp->primaryScreen()->size().width());
     htmlPreview->setObjectName("htmlpreview");
     htmlPreview->setVisible(appSettings->htmlPreviewVisible());
-
-    previewSplitter = new QSplitter(this);
-    previewSplitter->addWidget(editorPane);
-    previewSplitter->addWidget(htmlPreview);
-    previewSplitter->setCollapsible(0, true);
-    previewSplitter->setCollapsible(1, true);
-
 
     this->findReplace = new FindReplace(this->editor, this);
     statusBarWidgets.append(this->findReplace);
@@ -327,67 +311,71 @@ MainWindow::MainWindow(const QString &filePath, QWidget *parent)
 
     buildMenuBar();
     buildStatusBar();
-    
-    QVBoxLayout *mainLayout = new QVBoxLayout();
-    QWidget *mainPane = new QWidget(this);
-    mainPane->setLayout(mainLayout);
-    mainLayout->addWidget(previewSplitter, 500);
-    mainLayout->setSpacing(0);
-    mainLayout->setMargin(0);
 
-    sidebarSplitter = new QSplitter(this);
-    sidebarSplitter->addWidget(sidebar);
-    sidebarSplitter->addWidget(mainPane);
-    sidebarSplitter->setCollapsible(0, false);
-    sidebarSplitter->setCollapsible(1, true);
     sidebar->setMinimumWidth(0.1 * qApp->primaryScreen()->size().width());
-    sidebar->setMaximumWidth(0.5 * this->width());
 
-    // If previous sidebar width was stored, reload it.
-    if (windowSettings.contains(GW_SIDEBAR_STATE_KEY)) {
-        sidebarSplitter->restoreState(windowSettings.value(GW_SIDEBAR_STATE_KEY).toByteArray());
-    }
-    // Else set default size of sidebar.
-    else {
-        QList<int> sidebarSplitterSizes;
+    splitter = new QSplitter(this);
+    splitter->addWidget(sidebar);
+    splitter->addWidget(editorPane);
+    splitter->addWidget(htmlPreview);
+    splitter->setChildrenCollapsible(false);
+    splitter->setStretchFactor(0, 0);
+    splitter->setStretchFactor(1, 2);
+    splitter->setStretchFactor(2, 1);
 
-        sidebarSplitterSizes.append(0.25 * this->width());
-        sidebarSplitterSizes.append(0.75 * this->width());
-        sidebarSplitter->setSizes(sidebarSplitterSizes);
-    }
+    // Set default sizes for splitter.
+    QList<int> sizes;
+    int sidebarWidth = this->width() * 0.2;
+    int otherWidth = (this->width() - sidebarWidth) / 2;
+    sizes.append(sidebarWidth);
+    sizes.append(otherWidth);
+    sizes.append(otherWidth);
 
-    this->setCentralWidget(sidebarSplitter);
+    splitter->setSizes(sizes);
 
-    // Show the main window.
-    show();
-
-    // Apply the theme only after show() is called on all the widgets,
-    // since the Outline scrollbars can end up transparent in Windows if
-    // the theme is applied before show().
-    //
-    applyTheme();
-    adjustEditorWidth(this->width(), true);
-
-    this->update();
-    qApp->processEvents();
-
-    if (!fileToOpen.isNull() && !fileToOpen.isEmpty()) {
-        documentManager->open(fileToOpen);
+    // If previous splitter geometry was stored, load it.
+    if (windowSettings.contains(GW_SPLITTER_GEOMETRY_KEY)) {
+        this->splitter->restoreState(windowSettings.value(GW_SPLITTER_GEOMETRY_KEY).toByteArray());
     }
 
-    if (fileLoadError) {
-        QMessageBox::critical
-        (
-            this,
-            QApplication::applicationName(),
-            tr("Could not create file %1. Check permissions.").arg(filePath)
-        );
-    }
+    this->connect(splitter,
+        &QSplitter::splitterMoved,
+        [this](int pos, int index) {
+            Q_UNUSED(pos)
+            Q_UNUSED(index)
+            adjustEditor();
+        });
+
+    this->setCentralWidget(splitter);
+
+    qApp->installEventFilter(this);
 
     toggleHideMenuBarInFullScreen(appSettings->hideMenuBarInFullScreenEnabled());
     menuBarMenuActivated = false;
 
-    qApp->installEventFilter(this);
+    // Need this call for GTK / Gnome 42 segmentation fault workaround.
+    qApp->processEvents();
+
+    show();
+
+    // Apply the theme only after show() is called on all the widgets,
+    // since the Outline scrollbars can end up transparent in Windows if
+    // the theme is applied before show().  We cannot call show() and
+    // then apply the theme in the constructor due to a bug with
+    // Wayland + GTK that causes a segmentation fault.
+    //
+    applyTheme();
+    adjustEditor();
+
+    // Show the theme right away before loading any files.
+    qApp->processEvents();
+
+    // Load file from command line or last session.
+    if (!fileToOpen.isNull() && !fileToOpen.isEmpty()) {
+        documentManager->open(fileToOpen);
+    }
+
+    spelling->startLiveSpellCheck();
 }
 
 MainWindow::~MainWindow()
@@ -402,7 +390,27 @@ QSize MainWindow::sizeHint() const
 
 void MainWindow::resizeEvent(QResizeEvent *event)
 {
-    adjustEditorWidth(event->size().width(), true);
+    int width = event->size().width();
+
+    if (width < (0.5 * qApp->primaryScreen()->size().width())) {
+        this->sidebar->setVisible(false);
+        this->sidebar->setAutoHideEnabled(true);
+        this->sidebarHiddenForResize = true;
+    }
+    else {
+        this->sidebarHiddenForResize = false;
+
+        if (!this->focusModeEnabled && this->appSettings->sidebarVisible()) {
+            this->sidebar->setAutoHideEnabled(false);
+            this->sidebar->setVisible(true);
+        }
+        else {
+            this->sidebar->setAutoHideEnabled(true);
+            this->sidebar->setVisible(false);
+        }
+    }
+
+    adjustEditor();
 }
 
 void MainWindow::keyPressEvent(QKeyEvent *e)
@@ -481,11 +489,11 @@ void MainWindow::quitApplication()
 
         windowSettings.setValue(GW_MAIN_WINDOW_GEOMETRY_KEY, saveGeometry());
         windowSettings.setValue(GW_MAIN_WINDOW_STATE_KEY, saveState());
-        windowSettings.setValue(GW_SIDEBAR_STATE_KEY, sidebarSplitter->saveState());
+        windowSettings.setValue(GW_SPLITTER_GEOMETRY_KEY, splitter->saveState());
         windowSettings.sync();
 
-        DictionaryManager::instance().addProviders();
-        DictionaryManager::instance().setDefaultLanguage(language);
+        DictionaryManager::instance()->addProviders();
+        DictionaryManager::instance()->setDefaultLanguage(language);
 
         this->editor->document()->disconnect();
         this->editor->disconnect();
@@ -523,17 +531,16 @@ void MainWindow::openPreferencesDialog()
 
 void MainWindow::toggleHtmlPreview(bool checked)
 {
-    QList<int> splitterSizes;
-
     htmlPreviewMenuAction->blockSignals(true);
 
     htmlPreviewMenuAction->setChecked(checked);
     htmlPreview->setVisible(checked);
     htmlPreview->updatePreview();
-    adjustEditorWidth(this->width());
 
     htmlPreviewMenuAction->blockSignals(false);
     appSettings->setHtmlPreviewVisible(checked);
+    this->update();
+    adjustEditor();
 }
 
 void MainWindow::toggleHemingwayMode(bool checked)
@@ -547,13 +554,20 @@ void MainWindow::toggleHemingwayMode(bool checked)
 
 void MainWindow::toggleFocusMode(bool checked)
 {
+    this->focusModeEnabled = checked;
+
     if (checked) {
-        editor->setFocusMode(appSettings->focusMode());        
+        editor->setFocusMode(appSettings->focusMode());
+        sidebar->setVisible(false);
+        sidebar->setAutoHideEnabled(true);
     } else {
         editor->setFocusMode(FocusModeDisabled);
-    }
 
-    sidebar->setAutoHideEnabled(checked);
+        if (!this->sidebarHiddenForResize && this->appSettings->sidebarVisible()) {
+            sidebar->setAutoHideEnabled(false);
+            sidebar->setVisible(true);
+        }
+    }
 }
 
 void MainWindow::toggleFullScreen(bool checked)
@@ -642,7 +656,7 @@ void MainWindow::toggleDisplayTimeInFullScreen(bool checked)
 void MainWindow::changeEditorWidth(EditorWidth editorWidth)
 {
     editor->setEditorWidth(editorWidth);
-    adjustEditorWidth(this->width());
+    adjustEditor();
 }
 
 void MainWindow::changeInterfaceStyle(InterfaceStyle style)
@@ -914,8 +928,6 @@ void MainWindow::onAboutToShowMenuBarMenu()
 
 void MainWindow::onSidebarVisibilityChanged(bool visible)
 {
-    this->adjustEditorWidth(this->width());
-
     if (visible) {
         toggleSidebarButton->setText(QChar(fa::chevronleft));
     } else {
@@ -925,17 +937,30 @@ void MainWindow::onSidebarVisibilityChanged(bool visible)
     if (!visible) {
         editor->setFocus();
     }
+
+    this->showSidebarAction->blockSignals(true);
+    this->showSidebarAction->setChecked(visible);
+    this->showSidebarAction->blockSignals(false);
+
+    this->adjustEditor();
 }
 
 void MainWindow::toggleSidebarVisible(bool visible)
 {
     this->appSettings->setSidebarVisible(visible);
-    this->sidebar->setAutoHideEnabled(!visible);
-    this->sidebar->setVisible(visible);
 
-    this->showSidebarAction->blockSignals(true);
-    this->showSidebarAction->setChecked(visible);
-    this->showSidebarAction->blockSignals(false);
+    if (!this->sidebarHiddenForResize
+            && !this->focusModeEnabled
+            && this->appSettings->sidebarVisible()) {
+        sidebar->setAutoHideEnabled(false);
+    }
+    else {
+        sidebar->setAutoHideEnabled(true);
+    }
+
+    this->sidebar->setVisible(visible);
+    this->sidebar->setFocus();
+    adjustEditor();
 }
 
 QAction* MainWindow::createWindowAction
@@ -1024,7 +1049,7 @@ void MainWindow::buildMenuBar()
     editMenu->addAction(createWindowAction(tr("Find &Next"), findReplace, SLOT(findNext()), QKeySequence::FindNext));
     editMenu->addAction(createWindowAction(tr("Find &Previous"), findReplace, SLOT(findPrevious()), QKeySequence::FindPrevious));
     editMenu->addSeparator();
-    editMenu->addAction(createWindowAction(tr("&Spell check"), editor, SLOT(runSpellChecker())));
+    editMenu->addAction(createWindowAction(tr("&Spell check"), spelling, SLOT(runSpellCheck())));
 
     QMenu *formatMenu = this->menuBar()->addMenu(tr("For&mat"));
     formatMenu->addAction(createWidgetAction(tr("&Bold"), editor, SLOT(bold()), QKeySequence::Bold));
@@ -1164,13 +1189,13 @@ void MainWindow::buildStatusBar()
 
     QHBoxLayout *leftLayout = new QHBoxLayout(leftWidget);
     leftWidget->setLayout(leftLayout);
-    leftLayout->setMargin(0);
+    leftLayout->setContentsMargins(0,0,0,0);
     QHBoxLayout *midLayout = new QHBoxLayout(midWidget);
     midWidget->setLayout(midLayout);
-    midLayout->setMargin(0);
+    midLayout->setContentsMargins(0,0,0,0);
     QHBoxLayout *rightLayout = new QHBoxLayout(rightWidget);
     rightWidget->setLayout(rightLayout);
-    rightLayout->setMargin(0);
+    rightLayout->setContentsMargins(0,0,0,0);
 
     // Add left-most widgets to status bar.
     QFont buttonFont(this->awesome->font(style::stfas, 16));
@@ -1328,8 +1353,9 @@ void MainWindow::buildSidebar()
     
     connect(this->showSidebarAction,
         &QAction::toggled,
-        this,
-        &MainWindow::toggleSidebarVisible);
+        [this](bool visible) {
+            this->toggleSidebarVisible(visible);
+        });
 
     cheatSheetWidget = new QListWidget(this);
 
@@ -1401,8 +1427,9 @@ void MainWindow::buildSidebar()
     connect(editor, SIGNAL(typingResumed()), sessionStats, SLOT(onTypingResumed()));
 
     sidebar = new Sidebar(this);
-    sidebar->setMinimumWidth(0.10 * QGuiApplication::primaryScreen()->availableSize().width());
-    sidebar->setMaximumWidth(0.33 * QGuiApplication::primaryScreen()->availableSize().width());
+    sidebar->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Preferred);
+    sidebar->setMinimumWidth(0.1 * QGuiApplication::primaryScreen()->availableSize().width());
+    sidebar->setMaximumWidth(0.5 * QGuiApplication::primaryScreen()->availableSize().width());
 
     QPushButton *tabButton = new QPushButton();
     tabButton->setFont(this->awesome->font(style::stfas, 16));
@@ -1476,52 +1503,41 @@ void MainWindow::buildSidebar()
         this,
         &MainWindow::onSidebarVisibilityChanged);
     
-    this->toggleSidebarVisible(appSettings->sidebarVisible());
+    if (!this->sidebarHiddenForResize
+            && !this->focusModeEnabled
+            && appSettings->sidebarVisible()) {
+        this->sidebar->setAutoHideEnabled(false);
+        this->sidebar->setVisible(true);
+    }
+    else {
+        this->sidebar->setAutoHideEnabled(true);
+        this->sidebar->setVisible(false);
+    }
 }
 
-void MainWindow::adjustEditorWidth(int width, bool resizeEvent)
+void MainWindow::adjustEditor()
 {
-    static bool sidebarHiddenForResize = false;
-    static QList<int> prevSidebarSizes = sidebarSplitter->sizes();
+    // Make sure editor size is updated.
+    qApp->processEvents();
 
-    QList<int> previewSplitterSizes;
-    int editorWidth = width;
+    int width = this->width();
+    int sidebarWidth = 0;
 
-    if (resizeEvent) {
-        sidebar->setMaximumWidth(0.5 * this->width());
-
-        if (width < (0.5 * qApp->primaryScreen()->size().width())) {
-            if (!sidebarHiddenForResize) {
-                prevSidebarSizes = sidebarSplitter->sizes();
-            }
-
-            toggleSidebarVisible(false);
-            sidebarHiddenForResize = true;
-        }
-        else {
-            if (sidebarHiddenForResize) {
-                toggleSidebarVisible(true);
-                sidebarHiddenForResize = false;
-                sidebarSplitter->setSizes(prevSidebarSizes);
-            }
-
-            editorWidth -= sidebar->width();
-        }
+    // Make sure live preview does not crowd out editor.
+    // It should not take up more than 50% of the window space
+    // left after the sidebar is accounted for.
+    //
+    if (this->sidebar->isVisible()) {
+        sidebarWidth = this->sidebar->width();
     }
 
-    if (htmlPreview->isVisible()) {
-        editorWidth /= 2;
-        previewSplitterSizes.append(editorWidth);
-    }
+    this->htmlPreview->setMaximumWidth((width - sidebarWidth) / 2);
 
-    previewSplitterSizes.append(editorWidth);
-    previewSplitter->setSizes(previewSplitterSizes);
-
-    // Resize the editor's margins based on the size of the window.
-    editor->setupPaperMargins();
+    // Resize the editor's margins.
+    this->editor->setupPaperMargins();
 
     // Scroll to cursor position.
-    editor->centerCursor();
+    this->editor->centerCursor();
 }
 
 void MainWindow::applyTheme()
@@ -1543,6 +1559,7 @@ void MainWindow::applyTheme()
 
     editor->setColorScheme(colorScheme);
     editor->setStyleSheet(styler.editorStyleSheet());
+    spelling->setErrorColor(colorScheme.error);
 
     // Do not call this->setStyleSheet().  Calling it more than once in a run
     // (i.e., when changing a theme) causes a crash in Qt 5.11.  Instead,
@@ -1550,8 +1567,7 @@ void MainWindow::applyTheme()
     //
     qApp->setStyleSheet(styler.layoutStyleSheet());
 
-    previewSplitter->setStyleSheet(styler.splitterStyleSheet());
-    sidebarSplitter->setStyleSheet(styler.splitterStyleSheet());
+    this->splitter->setStyleSheet(styler.splitterStyleSheet());
     this->statusBar()->setStyleSheet(styler.statusBarStyleSheet());
 
     foreach (QWidget *w, statusBarWidgets) {
@@ -1575,7 +1591,7 @@ void MainWindow::applyTheme()
 
     htmlPreview->setStyleSheet(styler.htmlPreviewCss());
 
-    adjustEditorWidth(this->width());
+    adjustEditor();
 }
 
 } // namespace ghostwriter